requirements.txt
# Core
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.19

# APIs
requests==2.31.0
schedule==1.2.0

# Analysis
scikit-learn==1.3.0
xgboost==1.7.6
prophet==1.1.4

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Jupyter
jupyter==1.0.0
notebook==7.0.2
